+++ CHANGELOG.md
@@ -4,9 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ## [3.2.0] - TBD
 ### Added
-- `apiary-metastore-events` module that contains `apiary-metastore-listener` and `apiary-recivers`
+- `apiary-metastore-events` module that contains `apiary-metastore-listener` and `apiary-receivers`
 ### Changed
-- `apiary-metastore-listener` and `apiary-recivers` package paths were renamed from: `com.expediagroup.apiary.extensions` to: `com.expediagroup.apiary.extensions.events`
+- `apiary-metastore-listener` and `apiary-receivers` package paths were renamed from `com.expediagroup.apiary.extensions` to `com.expediagroup.apiary.extensions.events`
 
 ## [3.1.0] - 2019-05-03
 ### Fixed